+++ tests/test_onyx_analysis_helper.py
@@ -8,6 +8,7 @@ WARNING: Using --basetemp on an existing folder will overwrite all files.
 """
 
 import datetime
+import logging
 from pathlib import Path
 from unittest.mock import patch
 
@@ -724,8 +725,8 @@ def test_add_versions_to_methods_just_onyx(caplog):
 
 def test_add_versions_to_methods_plus_tools(caplog):
     """
-    Test that add_methods functions gets the versions from the query when set to true and adds
-    user defined tool versions and then populates the attribute.
+    Test that add_methods functions gets the versions from the query (plus onyx versions hash)
+    and adds user defined tool versions and then populates the methods attribute correctly.
     """
 
     expected_results = {
@@ -737,7 +738,8 @@ def test_add_versions_to_methods_plus_tools(caplog):
             {"name": "sylph_db_version", "version": "1.0.0"},
             {"name": "alignment_db_version", "version": "1.0.0"},
             {"name": "my_pkg", "version": "v1.2.3"},
-        ]
+        ],
+        "onyx_versions_hash": "e0c8c12a02fa86494059858c41af311d94c086a286bf4c62d53c21261e90f614",
     }
 
     analysis = oa.OnyxAnalysis()
@@ -753,14 +755,15 @@ def test_add_versions_to_methods_plus_tools(caplog):
 
 def test_add_versions_do_not_overwrite(caplog):
     """
-    Test that adding versions doesn't overwrite
+    Test that adding versions doesn't overwrite. Includes onyx versions and hash.
     """
     expected_methods = {
         "versions": [
             {"name": "classifier_version", "version": "1.0.0"},
             {"name": "my_pkg", "version": "v1.2.3"},
             {"name": "my_other_pkg", "version": "v2.3.4"},
-        ]
+        ],
+        "onyx_versions_hash": "b997b78b7ef8c0e21d8d6c0fe242bb9f5e0b98ff6e13240ef4e341d787605481",
     }
     analysis = oa.OnyxAnalysis()
     methods_fail = analysis.add_versions_to_methods(
@@ -776,8 +779,10 @@ def test_add_versions_do_not_overwrite(caplog):
         tool_versions={"my_other_pkg": "v2.3.4"},
     )
     assert not methods_fail_3
-    assert analysis.methods == expected_methods
-    print(analysis.methods)
+    assert analysis.methods == expected_methods, (
+        f"Actual methods attribute does not look as expected: {expected_methods}"
+    )
+    print(f"Expected methods to correct look like: {analysis.methods}")
 
 
 def test_add_versions_to_methods_onyx_versions_not_list(caplog):
@@ -1067,3 +1072,150 @@ def test_add_methods_broken_methods_dict_input(caplog):
     assert "Error: Methods must be in dict format." in caplog.text
     assert methods_fail
     print("Error correctly caught when input type not dict.")
+
+
+@pytest.mark.parametrize(
+    "vers,truncate,expect",
+    [
+        ("1.2.3", "MAJOR", "1"),
+        ("1.2.3", "MINOR", "1.2"),
+        ("1.2.3", "PATCH", "1.2.3"),
+        ("1.2.3-rc.4", "MINOR", "1.2"),
+        ("1.2.3-rc.4", "PATCH", "1.2.3"),
+        ("not a version", "MAJOR", "not a version"),
+    ],
+)
+def test_truncate_version(vers, truncate, expect):
+    actual = oa.truncate_version(vers, truncate)
+    assert actual == expect
+
+
+MOCK_ANALYSIS_RECORD = [
+    {
+        "published_date": "1970-01-01",
+        "site": "test",
+        "analysis_id": "AID-12345678",
+        "analysis_date": "1970-01-01",
+        "name": "test-analysis",
+        "report": "",
+        "outputs": "path/to/outputs/file.json",
+    }
+]
+
+MOCK_ANALYSIS_TABLE = {
+    "name": "test-analysis",
+    "description": "This is a test analysis",
+    "analysis_date": "1970-01-01",
+    "pipeline_name": "test-pipeline",
+    "pipeline_url": "test-pipeline-url",
+    "pipeline_version": "0.1.0",
+    "result": "test result",
+    "upstream_analyses": [],
+    "report": "",
+    "outputs": "path/to/outputs/file.json",
+    "methods": {
+        "versions": [
+            {"name": "a_great_tool", "version": "1.0.0"},
+            {"name": "another_great_tool", "version": "2000.0.0"},
+        ],
+        "thresholds": {"limit": 10},
+        "method2": "method example 2",
+    },
+    "result_metrics": {
+        "Example result 1": 9,
+        "Example result 2": "Fail",
+        "Example result 3": 0.3,
+    },
+    "synthscape_records": ["ID-123456789"],
+    "identifiers": [],
+    "analysis_id": "AID-12345678",
+}
+
+ANOTHER_MOCK_ANALYSIS_RECORD = [
+    {
+        "published_date": "1970-01-02",
+        "site": "test-the-second",
+        "analysis_id": "AID-89012345",
+        "analysis_date": "1970-01-02",
+        "name": "test-analysis",
+        "report": "",
+        "outputs": "path/to/file_2.json",
+    }
+]
+
+ANOTHER_MOCK_ANALYSIS_TABLE = {
+    "name": "test-analysis",
+    "description": "This is another test analysis",
+    "analysis_date": "1970-01-02",
+    "pipeline_name": "test-pipeline",
+    "pipeline_url": "test-pipeline-url",
+    "pipeline_version": "0.1.0",
+    "result": "another test result",
+    "upstream_analyses": [],
+    "report": "",
+    "outputs": "path/to/file_2.json",
+    "methods": {
+        "versions": [
+            {"name": "a_great_tool", "version": "1.0.0"},
+            {"name": "another_great_tool", "version": "2000.0.0"},
+        ],
+        "thresholds": {"limit": 10},
+        "method2": "method example 2",
+    },
+    "result_metrics": {
+        "Example result 1": 9,
+        "Example result 2": "Fail",
+        "Example result 3": 0.3,
+    },
+    "synthscape_records": ["ID-123456789"],
+    "identifiers": [],
+    "analysis_id": "AID-89012345",
+}
+
+
+@patch(
+    target="onyx_analysis_helper.onyx_analysis_helper_functions.OnyxClient.get_analysis",
+    return_value=MOCK_ANALYSIS_TABLE.copy(),
+)
+@patch(
+    target="onyx_analysis_helper.onyx_analysis_helper_functions.OnyxClient.analyses",
+    return_value=MOCK_ANALYSIS_RECORD.copy(),
+)
+def test_get_analysis_records(mocked_analyses, mocked_analysis_table):
+    analyses_records, exitcode = oa.get_analysis_records(sample_id="ID-123456", server="")
+    assert len(analyses_records) == 1
+    assert exitcode == 0
+
+
+@patch(
+    "onyx_analysis_helper.onyx_analysis_helper_functions.OnyxClient.analyses",
+)
+def test_get_analysis_records_multiple_analyses(mocked_analyses):
+    """
+    Get analyses tables from sample with multiple analyses tables. Mock the get_analysis records and
+    the analysis table records.
+
+    Note that this function sometimes fails but running alone seems to pass?
+    """
+    mocked_analyses.return_value = MOCK_ANALYSIS_RECORD + ANOTHER_MOCK_ANALYSIS_RECORD
+
+    with patch(
+        "onyx_analysis_helper.onyx_analysis_helper_functions.OnyxClient.get_analysis",
+        side_effect=[MOCK_ANALYSIS_TABLE.copy(), ANOTHER_MOCK_ANALYSIS_TABLE.copy()],
+    ):
+        many_analyses_records, exitcode = oa.get_analysis_records(sample_id="ID-123456", server="")
+        print(many_analyses_records)
+        assert len(many_analyses_records) == 2
+        assert exitcode == 0
+
+
+@patch("onyx_analysis_helper.onyx_analysis_helper_functions.OnyxClient.analyses")
+def test_get_analysis_records_no_analyses(mocked_analyses, caplog):
+    caplog.set_level(logging.INFO)
+
+    mocked_analyses.return_value = []
+
+    analyses_records, exitcode = oa.get_analysis_records(sample_id="ID-123456", server="")
+    assert "No analysis tables found for sample ID-123456" in caplog.text
+    assert analyses_records == {}
+    assert exitcode == 0